combine some files
